Player rebalance + bug fix
player no longer gets stuck in jitter loop - in the process of doing this I also made the player exhibit different behavior, entering stationary a short while BEFORE actually being completely stationary. Feels more responsive to the player and gives a cool but subtle drift to the worm when coming to a stop to shoot.

Unfortunately this broke animations a bit because timing is all different now (much more consistent now, but with different timing)